moved profiler interface to dataset module
